Start battle scene
